A big push getting ready for the deployment.  Most of these fixes were dealing with user space issues, like workspaces and logging.

+++ graphene-parent/graphene-dao/src/main/java/graphene/services/LinkGenerator.java
@@ -5,6 +5,6 @@ import org.apache.tapestry5.Link;
 public interface LinkGenerator {
 
 	public abstract Link set(final String schema, final String type, final String match, final String value,
-			final int maxResults);
+			final long maxResults);
 
 }
